src/04_data_ingestion/requirements.txt
snowflake-snowpark-python
yfinance  
boto3 
